% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the parameter p that maximizes the log-likelihood for Bernoulli distributed data}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A numeric vector of 0s and 1s.}
}
\value{
Numeric value of p that maximizes the log-likelihood.
}
\description{
This function takes a numeric vector consisting of 0s and 1s and calculates the value of p
that maximizes the log-likelihood of a Bernoulli distribution.
}
\examples{
logLikBernoulli(c(1, 0, 0, 0, 1, 1, 1))
}
