% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate the minimum sample size needed for a t-test with specified power and significance level}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{Numeric vector representing the first sample.}

\item{x2}{Optional numeric vector representing the second sample.}
}
\value{
Minimum sample size needed for the specified statistical power.
}
\description{
This function calculates the minimum number of observations needed to achieve a specified power
and significance level for a one-sample or two-sample t-test, depending on whether one or two
samples are provided.
}
\examples{
minimumN(rnorm(10, mean = 0, sd = 1))
}
