% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Returns an approximation to the data based on the specified number of principal components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{Numeric matrix or data frame.}

\item{npc}{Integer indicating the number of principal components to use.}
}
\value{
Numeric matrix approximating the original data.
}
\description{
This function performs PCA on the provided data matrix and returns an approximation of the
original data based on the specified number of principal components.
}
\examples{
pcApprox(matrix(rnorm(100), ncol=10), npc=2)
}
