% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names to Camel Case}
\usage{
standardizeNames(data, case_type = "small_camel")
}
\arguments{
\item{data}{A tibble whose variable names need to be standardized.}

\item{case_type}{The type of case to convert variable names into.
Default is 'small_camel'. Other options like 'big_camel', 'snake', etc.,
can be found in the documentation of the \code{snakecase::to_any_case} function.}
}
\value{
A tibble with standardized variable names.
}
\description{
A function to standardize variable names in a tibble to small camel case or
another specified case using \code{dplyr::rename_with} and \code{janitor::make_clean_names}.
The function relies on the \code{snakecase} package to format names into the chosen case.
}
\examples{
df <- tibble::tibble(`Old Name1` = 1:10, `another name2` = letters[1:10])
new_df <- standardizeNames(df, 'small_camel')
}
