% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Calculate and plot a survival curve S(t) using given status and time data}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numeric vector indicating the occurrence of an event (1) or censoring (0).}

\item{time}{Numeric vector indicating the time to event or censoring.}
}
\value{
A ggplot object representing the survival curve.
}
\description{
This function takes two numeric vectors: 'status' indicating the event occurrence and 'time'
indicating the time until the event or censoring occurs. It calculates and plots the survival
curve S(t).
}
\examples{
survCurv(status = c(1, 0, 1), time = c(5, 10, 15))
}
