% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Reverse the centering and scaling transformation applied to a vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A scaled numeric vector.}
}
\value{
Numeric vector with reversed scaling and centering transformations.
}
\description{
This function takes a scaled vector and reverses the scaling and centering transformations
applied to it, returning the original vector.
}
\examples{
x <- scale(1:10)
unscale(x)
}
